+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Use `Hooks` registry class instead of globals for hooks
+- `reify_constraints` hook is called
+- `reify_var` hook is used instead of unconditionally creating `ReifiedVar`
+
+### Added
+- `reify_value` hook
+
 ## [0.2.2] - 2023-04-08
 
 ### Fixed